[feat] 프로젝트 생성기능 수정 - 프로젝트 생성에서 이미지 업로드 시 이미지 엔티티 및 프로젝트 엔티티의 매핑기능추가

+++ src/main/java/com/team1/TBBCloneCoding/project/dto/ProjectCreateRequestDto.java
@@ -16,5 +16,5 @@ public class ProjectCreateRequestDto {
     private Long goalPrice;
     private LocalDateTime statDate;
     private LocalDateTime endDate;
-    private List<Integer> imageList;
+    private List<Long> imageList;
 }
